Fix wave restarts and GEFS FHOUT/FHMAX (#3009)
# Description
Fixes some issues that were preventing wave restarts from operating
correctly.

First, the wave restart files were not being correctly linked from
`$DATA` to `$DATArestart`. The files are placed in the root of `$DATA`
instead of in `${DATA}/WAVE_RESTART`, so now links for the individual
files are created.

Second, the incorrect filenames were being searches for and copied as
part of a rerun. Filenames were geared towards multigrid waves, which
use the grid names, but single grid just uses a `ww3`. Since multigrid
waves are deprecated in workflow and will soon be removed (#2637), these
were updated only supporting the single-grid option.

These fixes allow forecast segments (and emergency restarts) to work
correctly when waves are on.

Additionally, the `FHMAX` (and `FHOUT`) for perturbed GEFS members was
being overwritten by `config.efcs` due to remnant code from when it was
copied from the EnKF version. This interfered with the segment setting
for those GEFS members.

Resolves #3001

# Type of change
- [x] Bug fix (fixes something broken)
- [ ] New feature (adds functionality)
- [ ] Maintenance (code refactor, clean-up, new CI test, etc.)

# Change characteristics
- Is this a breaking change (a change in existing functionality)? NO
- Does this change require a documentation update? NO
- Does this change require an update to any of the following submodules?
NO

# How has this been tested?
- S2SW forecast-only test with segments on Hercules

# Checklist
- [x] Any dependent changes have been merged and published
- [x] My code follows the style guidelines of this project
- [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas
- [x] I have documented my code, including function, input, and output
descriptions
- [x] My changes generate no new warnings
- [x] New and existing tests pass with my changes
- [x] This change is covered by an existing CI test or a new one has
been added
- [x] I have made corresponding changes to the system documentation if
necessary

---------

Co-authored-by: Kate.Friedman <Kate.Friedman@noaa.gov>
Co-authored-by: David Huber <69919478+DavidHuber-NOAA@users.noreply.github.com>