Fix wave restart for cold start and add ic version file (#3112)
The stage job was incorrectly putting wave restarts into the gfs
directory. The forecast job looks for them in the gdas directory, so
this is updated.

Additionally, the restarts were also not being copied from the staged
directory to `$DATA`, so now they are. The process is identical to that
of non-RERUN warm starts, so the code is refactored a bit to avoid
duplication.

As part of updating the ICs with the new restart location, an IC version
file is added to support different version numbers for different IC
directories. Unlike other version files, this one uses an associative
array rather than individual variables.

Resolves #3109