Use stochastic restart patterns on rerun (#3077)
The stochastic pattern restart files were not being copied into the
input directory when restarting the model after a segment/failure. These
files are now copied in and the stochini flag set to `.true.` in the
namelist on a rerun.
    
The files are NOT copied in for non-rerun warm starts.

Also removes the restriction that stochastic physics cannot be run on
member 0, as this is desired down the line. Additional settings are
added to the fcst and efcs configs to retain the current behavior.

A bug was also discovered and corrected during this work. The stage,
forecast, and archive job all assumed that `ca_data` tile files are
always present, but these are only created when cellular automata is on.
Now `ca_data` files are handled if CA is on. To implement this change,
the `DO_CA` setting had to be moved from the forecast configs to base so
it is available to the stage_ic and archive jobs.

Resolves #2937