Allow building of the ufs-weather-model, WW3 pre/post execs for GFS, GEFS, SFS in the same clone of global-workflow (#3098)
GFS and GEFS (and now SFS) uses different compile time options for the
UFS-weather-model. For the purposes of CI testing, a multi-build
pipeline under Jenkins was created by @TerrenceMcGuinness-NOAA. This
served well, until now. With the inclusion of SFS, a third variety of
the model is being built. Under the multi-build pipeline paradigm, a
second (or third) clone and build of the global-workflow is required.
This adds cloning and compilation time of the global-workflow.

This PR allows compiling the ufs-weather-model in a single clone of the
global-workflow. The compiled executables based on the options for GFS,
GEFS, and SFS results in a model executable as `gfs_model.x`,
`gefs_model.x`, and `sfs_model.x`. The forecast script uses the right
executable.

This PR also differentiates the WW3 pre/post executables based on `gfs_`
or `gefs_`. SFS variants need to be introduced, when `NET=sfs` is added.
In the process of updating the WW3 pre/post executable names, it was
discovered (via `grep`) the following are not used:
- `ww3_prep`
- `ww3_outf`
- `ww3_ounf`
- `ww3_ounp`
These need to be confirmed by running the workflow to ensure they are
indeed not used at runtime.

This PR also updates the `Jenkinsfile` to use the multi-build from the
single location.
---------

Co-authored-by: Walter Kolczynski - NOAA <Walter.Kolczynski@noaa.gov>
Co-authored-by: David Huber <69919478+DavidHuber-NOAA@users.noreply.github.com>