Change orog gravity wave drag scheme for grid sizes less than 10km (#3175)
It has been found that while the HR4 GFS update improves the forecast
skill compared to the HR3 update, its skill is still less than that of
the HR2 update. To further improve the HR4 forecast skill, the tunable
parameter in the OGWD, 'effective grid size', has been tested. The
magnitude of the OGWD is inversely proportional to the effective grid
size (dx). The current default value of the effective grid size (dx) is
6*dx, which was optimally tuned for the C768 resolution (13km). It
appears that the 6*dx of the effective dx leads to too small OGWD for
the C1152 resolution (9km). Therefore, for the higher resolution of
which grid sizes are less than 10 km (i.e., C1152 and C3072) the
effective dx has been reduced to 2*dx in this update, which increased
the OGWD and significantly improved the GFS forecast skill in the C1152
resolution, compatible to the HR2 forecast skill.

Resolves: #3181