Fix mod_icec bug in atmos_prod (#3167)
The purpose of this PR is to fix a bug that causes the ensstat task to
occasionally produce a segmentation fault error. This segmentation fault
error is due to undefined values not being skipped in the ensemble
spread and mean calculations in the ensstat program. The reason
undefined values were not being skipped is because variables with
undefined values were not using bitmap in the pgrb files. Ensstat
expects undefined variables to use bitmap, otherwise ensstat will not
skip those undefined values.

The undefined variables were not using bitmap because of a bug in the
atmos_prod task, where the mod_icec function was not being skipped for
grib2 files that did not contain the LAND and ICEC variables. In the
[offline
UPP](https://github.com/NOAA-EMC/UPP/blob/develop/ush/fv3gfs_dwn_nems.sh),
the mod_icec functionality was being executed when LAND and ICEC
variables existed in the grib2 file. This same condition has been
applied in this PR for mod_icec.

Resolves #3150