Separate use of initial increment/perturbation file from REPLAY/+03 ICs  (#3119)
This PR is a continuation of the previous SFS PR. The use of initial
perturbations/increments for ensemble long forecasts (e.g., SFS and
GEFS) was missing from the previous PR because this option was included
with the REPLAY_IC option.

In this PR, the REPLAY_IC option is now strictly for the +03 ICs, and
the variables USE_ATM_ENS_PERTURB_FILES and USE_OCN_ENS_PERTURB_FILES
are used for initialization with the initial perturbation/increment
files used for SFS and GEFS reforecasting. This separation will also
make it easier to remove the +03 IC optoin.

---------

Co-authored-by: Eric.Sinsky <eric.sinsky@noaa.gov>
Co-authored-by: Walter Kolczynski - NOAA <Walter.Kolczynski@noaa.gov>
Co-authored-by: Rahul Mahajan <aerorahul@users.noreply.github.com>
Co-authored-by: Xiaqiong.Zhou <Xiaqiong.Zhou@noaa.gov>