Adding a gefs_arch task to GEFS workflow (#3211)
The `gefs_arch` task is not generated during the execution of the
workflow setup scripts. It is only created when `do_extractvars` is set
to true, with `do_extractvars` being false by default. This pull request
ensures that the `gefs_arch` task will be created in the workflow
regardless of the value of `do_extractvars`.

Resolves #3151