Copy post-processed sea ice increment for diagnostics (#3235)
This change copies post-processed sea ice increment (the increment that
was added to CICE restart files, as opposed to the raw increment that
was produced by SOCA), so it can be used in the
verification/diagnostics.

It also cleans up previously committed code (removes a list of files
where it's not needed anymore).

Resolves https://github.com/NOAA-EMC/GDASApp/issues/1402

---------

Co-authored-by: shlyaeva <anna.v.shlyaeva@noaa.gov>
Co-authored-by: RussTreadon-NOAA <26926959+RussTreadon-NOAA@users.noreply.github.com>