Splitting up the archive task (#3242)
In this PR the archive task will be split into two parts. The first
will always run and will just be for the copying of verification data to
the `VRFY_ARC` and `ARCDIR` directories. The second will only run when
`HPSSARCH` or `LOCALARCH` is set to YES and will generate and store
tarballs in the `ATARDIR`, either on `HPSS` or locally.

Resolves #3152