Capture build fail in Jenkins pipeline when no error logs are produced (#3297)
The catch block for a build fail in the Jenkins pipeline did not capture
the case when there are no error logs produced.
This is has been rectified with this update. The compute build shell
script still could be enhanced by creating a `error.log` file in the log
directory.

Resolves #3296