Fix the empty aerosol DA aerostat tar file issue (#3332)
Somewhere along the way, an inconsistency between the aerosol DA
diagnostic file output name suffixes and the workflow crept in, where
the files were generated, but not found via glob and put in a tarball
because of `.nc` vs `.nc4`. This PR resolves this issue plus also
renames the file to be more descriptive, e.g.:
`gdas.t00z.aerostat.tgz`
to clarify it is a tar ball with gzip. The PR also makes it a tarball
with gzip instead of an uncompressed tar.

  Resolves #3327