Add missing arch_tars dependencies (#3319)
The `gfs_arch_tars` job currently does not depend on `gempak` jobs,
even though it archives data produced by them. This PR will introduce
that dependency. Additionally, there are several missing dependencies
for cleanup when the arch_tar job is not executed. Nearly all of the
job's dependencies need to be replicated for cleanup in case arch_tar
doesn't run. This PR will address this problem as well.

Resolves #3294