Remove Direct Linking to COM from DATA for `wavepostsbs` Job (#3303)
In the current GFS version, some jobs directly write to COMOUT via
links in the working directories. This approach is risky, as downstream
jobs might receive incomplete files and fail. In order to resolve this
issue direct linking to COM from DATA should be removed. This PR
addresses this issue in the `wavepostsbs` job.

Resolves #3302 
Refs #712 
Refs #3273
---------

Co-authored-by: Rahul Mahajan <aerorahul@users.noreply.github.com>