Allow for C1152 ATM-Aero cycled DA to run on WCOSS2 (#3309)
This PR enables the ability to cycle C1152 atm+aerosol on WCOSS2 to
support GFSv17.
Wallclock times were extended because 1) model execution is now slower
with GOCART (this should probably be configurable in the
config.resources depending on APP but that is outside the scope of my
PR) 2) the aerosol DA variational solver needs more time at this
resolution and 3) it takes FOREVER to copy C1152 restart files with the
GOCART aerosol species in them.

Also, the prep job runs out of memory so this spreads it out over more
nodes. There are also misc. things that were added to get things to run
on WCOSS2 that may need altered/removed post the most recent RFC.

Finally, GDAS hash is updated.

Resolves #3275

---------

Co-authored-by: CatherineThomas-NOAA <59020064+CatherineThomas-NOAA@users.noreply.github.com>
Co-authored-by: Rahul Mahajan <aerorahul@users.noreply.github.com>