Replace WAVECUR_DID variable with "rtofs" (#3337)
This PR completes the remaining requested change that resolves bugzilla
1227. Per NCO guidance in bugzilla 1227 discussion, this PR removes the
`WAVECUR_DID` variable and replaces it with hardcoded "rtofs". This
matches with how the system handles obsproc paths with the hardcoded
"obsproc" text.

The other aspects of bugzilla 1227 have been remedied through other PRs
that removed the old `COMROOTp2` references and renamed `COM` variables.

The waveprep job is currently disabled in `develop` so CI is moot.
However, a sanity check was done by running the C48_S2SW CI test on
Hera.

Resolves #1246